+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### v2.0.0 (11 September 2016)
 
 * [ENHANCEMENT] Add support for IE<=11 (thanks @le0nik) (#22, #23)
+* [ENHANCEMENT] Provide an `isLeaving` prop on the current child component while it is leaving. (thanks @le0nik)
 * [ENHANCEMENT] Add ability to provide custom `height` className (useful for `css-modules`) (thanks @le0nik) (#20, #21)
 * [BUGFIX] Fix multiple remounts of children (thanks @le0nik) (#24)
 * [BUGFIX] Allow the child, that is leaving, to be `null` (thanks @le0nik)